my_requirements.txt
absl-py==0.2.2
alembic==0.9.10
APScheduler==3.5.1
apturl==0.5.2
astor==0.6.2
backcall==0.1.0
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
Brlapi==0.6.4
chainer==4.2.0
chainerui==0.4.0
chardet==2.3.0
checkbox-support==0.22
click==6.7
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
decorator==4.3.0
defer==1.0.6
entrypoints==0.2.3
feedparser==5.1.3
filelock==3.0.4
Flask==1.0.2
gast==0.2.0
getch==1.0
grpcio==1.13.0
guacamole==0.9.2
h5py==2.8.0
html5lib==0.9999999
httplib2==0.9.1
idna==2.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
language-selector==0.1
llvmlite==0.23.2
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.2.2
mistune==0.8.3
msgpack==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numba==0.38.1
numpy==1.14.5
oauthlib==1.0.3
onboard==1.2.0
opencv-python==3.4.1.15
padme==1.1.1
pandocfilters==1.4.2
parso==0.3.0
pexpect==4.0.1
pi==0.1.2
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
plainbox==0.25
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.7.3
python-debian==0.1.27
python-editor==1.0.3
python-systemd==231
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
pyzmq==17.0.0
reportlab==3.3.0
requests==2.9.1
scikit-learn==0.19.2
scipy==1.1.0
screen-resolution-extra==0.0.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
SQLAlchemy==1.2.9
ssh-import-id==5.5
system-service==0.3
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
torch==0.4.0
torchvision==0.2.1
tornado==5.0.2
tqdm==4.23.4
traitlets==4.3.2
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
wcwidth==0.1.7
Werkzeug==0.14.1
WMI==1.4.9
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3

